Updated strategies to allow for new Custom deployment strategy
